Changes to NA_pyfits and NP_pyfits to support opening fits files with a colon in the file name
git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@241 ed100bfc-0583-0410-97f2-c26b58777a21

+++ lib/NP_pyfits.py
@@ -3976,7 +3976,7 @@ class _py_File:
         if mode not in _python_mode.keys():
             raise "Mode '%s' not recognized" % mode
 
-        if mode != 'append':
+        if mode != 'append' and not os.path.exists(name):
            self.name, fileheader = urllib.urlretrieve(name)
         else:
            self.name = name
